RF: fold assets.py into utils.py; alphabetize
These had little to do with an asset and are mostly helpers for fsck.
